"scenario and enemy render updated"

+++ test/unisadventures/se_project/model/basicObjects/CollectibleItemTest.java
@@ -15,10 +15,10 @@ import unisadventures.se_project.util.CollectibleType;
  */
 public class CollectibleItemTest {
     
-    private final double _xPosition;
-    private final double _yPosition;
-    private final double _height;
-    private final double _width;
+    private final int _xPosition;
+    private final int _yPosition;
+    private final int _height;
+    private final int _width;
     private final ArrayList<String> _imageFileNameList;
     private final CollectibleType _type;
