+++ test/unisadventures/se_project/model/basicObjects/DestructibleObjectTest.java
@@ -14,10 +14,10 @@ import static org.junit.jupiter.api.Assertions.*;
  */
 public class DestructibleObjectTest {
     
-    private final double _xPosition;
-    private final double _yPosition;
-    private final double _height;
-    private final double _width;
+    private final int _xPosition;
+    private final int _yPosition;
+    private final int _height;
+    private final int _width;
     private final ArrayList<String> _imageFileNameList;
     
     public DestructibleObjectTest() {
